# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MobiDuka Retail POS API Blueprint",
    "description": "Multi-tenant, offline-first production backend engine supporting Flutter mobile clients and Kenyan retail ecosystems (Cash tracking, M-Pesa records, local synchronization syncing).",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "Current app origin API router"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT issued by /auth/login. Use the token as Authorization: Bearer <token>."
      }
    }
  },
  "security": [
    { "bearerAuth": [] }
  ],
...
  "paths": {
...
    "/customers": {
      "get": {
        "tags": ["Customers"],
        "summary": "Query or Search Registered Customer Profiles",
        "description": "Returns customers for a tenant, optionally filtered by name or phone number. The query filter is case-insensitive.",
        "parameters": [
          { "name": "businessId", "in": "query", "required": true, "schema": { "type": "string", "format": "uuid", "example": "0e722844-15f4-4807-8a69-9614e448f21e" }, "description": "Seeded deployed business ID." },
          { "name": "query", "in": "query", "required": false, "schema": { "type": "string" }, "description": "Case-insensitive name or phone search." }
        ],
        "responses": {
          "200": { "description": "Customer profiles returned successfully." },
          "400": { "description": "businessId is missing." }
        }
      },
      "post": {
        "tags": ["Customers"],
        "summary": "Register or Update Customer Profile",
        "description": "Creates or updates a customer and ensures an associated CreditAccount exists.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["businessId", "name"],
                "properties": {
                  "id": { "type": "string" },
                  "businessId": { "type": "string", "format": "uuid", "example": "0e722844-15f4-4807-8a69-9614e448f21e" },
                  "name": { "type": "string", "example": "Wanjiku Njoroge" },
                  "phone": { "type": "string", "example": "0722000111" },
                  "initialCreditLimit": { "type": "number", "minimum": 0, "example": 5000 }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Customer profile registered successfully." },
          "400": { "description": "Invalid profile data or duplicate phone number." }
        }
      }
    },
...
    "/employees": {
      "get": {
        "tags": ["Employees"],
        "summary": "Retrieve Multi-Tenant Staff Directory Roster",
        "parameters": [{ "name": "businessId", "in": "query", "required": true, "schema": { "type": "string", "format": "uuid", "example": "0e722844-15f4-4807-8a69-9614e448f21e" } }],
        "responses": {
          "200": { "description": "Roster array generated successfully." },
          "400": { "description": "businessId is missing." }
        }
      },
      "post": {
        "tags": ["Employees"],
        "summary": "Provision or Update a Staff Employee Account",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["businessId", "name", "role"],
                "properties": {
                  "id": { "type": "string" },
                  "businessId": { "type": "string", "format": "uuid", "example": "0e722844-15f4-4807-8a69-9614e448f21e" },
                  "name": { "type": "string", "example": "Kamau Mwangi" },
                  "email": { "type": "string", "format": "email" },
                  "phone": { "type": "string" },
                  "role": { "type": "string", "enum": ["OWNER", "MANAGER", "CASHIER"] },
                  "pin": { "type": "string", "example": "2541" },
                  "isActive": { "type": "boolean", "default": true }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Staff employee profile saved." },
          "400": { "description": "Invalid role, tenant, or employee configuration." }
        }
      }
    },
...
    "/expenses": {
      "get": {
        "tags": ["Business"],
        "summary": "Fetch Store Expense Ledger History",
        "description": "Returns tenant-scoped operational outflows ordered from newest to oldest.",
        "parameters": [
          { "name": "businessId", "in": "query", "required": true, "schema": { "type": "string", "format": "uuid", "example": "0e722844-15f4-4807-8a69-9614e448f21e" } }
        ],
        "responses": {
          "200": { "description": "Expense records fetched successfully." },
          "400": { "description": "businessId is missing." }
        }
      },
      "post": {
        "tags": ["Business"],
        "summary": "Log a New Cash or M-Pesa Store Outflow",
        "description": "Creates an expense record and an associated audit entry for the tenant user.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["businessId", "userId", "category", "amount", "description"],
                "properties": {
                  "id": { "type": "string" },
                  "businessId": { "type": "string", "format": "uuid", "example": "0e722844-15f4-4807-8a69-9614e448f21e" },
                  "userId": { "type": "string", "format": "uuid", "example": "33d0457d-ecb4-4e60-a05a-223bae2541a4" },
                  "category": { "type": "string", "example": "RENT" },
                  "amount": { "type": "number", "minimum": 0.01, "example": 15000 },
                  "description": { "type": "string", "example": "September shop rent payment" },
                  "paidTo": { "type": "string", "example": "Landlord" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Expense logged successfully." },
          "400": { "description": "Invalid expense data." },
          "404": { "description": "User was not found in the requested business." }
        }
      }
    }
  }
}